Revert "treewide: Bump rust nightly to 2023-03-30"
This reverts commit 6df95904607ddae8599f0a3c2e3fd4c67d7e751e.

Reason for revert: This is breaking the main build due to a bug when
running rustdoc

Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3714
Tested-by: Buildkite CI

+++ dataflow-expression/src/eval/builtins.rs
@@ -150,7 +150,7 @@ where
             .unwrap()
             .with_month(1)
             .unwrap()
-            .num_days_from_ce();
+            .num_days_from_ce() as i32;
         let mut weekday = if monday_first {
             time.weekday().num_days_from_monday()
         } else {
@@ -164,8 +164,8 @@ where
             }
             week_year = true;
             year -= 1;
-            days = days_in_year(year);
-            first_daynr -= days;
+            days = days_in_year(year) as i32;
+            first_daynr -= days as i32;
             weekday = (weekday + 53 * 7 - days) % 7;
         }
